Refactor Enum and Dict helpers
There were a few funcs shared between models that just smashed the
resource data regardless of type. This could create issues in the
future.

Trying something a little different to place serializations closer
to the resources that need them.

It introduces a fair bit of additional complexity, so not sure if
this is valuable or not.