Bug: Command.changes str used as dict in harness
The command execution harness was just passing an entire command object
into the operation implementations, which was a bug because the changes
were str rather than dict in that case. In the process is became clear
that passing all of the command was unnecessary and that passing in the
changes sub-dictionary would make more sense. Updated command execution
to transform changes block into kwargs for easier processing throughout
different operations. This is done in the harness rather than within
each operation.

Test data in test_api was updated to be correctly formed json, using
double quote instead of single.

FakeClient was unable to handle a conventional (error) payload from
the service. It was updated to handle error payloads as well as
conventional ones (errors).

Several additional tests were introduced to address gaps.